requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
asyncpg==0.29.0
redis[asyncio]==5.0.8
google-generativeai==0.8.3
APScheduler==3.10.4
pydantic==2.9.2
pydantic[email]==2.9.2
email-validator==2.2.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyJWT==2.9.0
bcrypt==4.2.0
